+++ azalea/examples/testbot/commands.rs
@@ -19,12 +19,13 @@ pub struct CommandSource {
 }
 
 impl CommandSource {
-    pub fn reply(&self, message: &str) {
+    pub fn reply(&self, message: impl Into<String>) {
+        let message = message.into();
         if self.chat.is_whisper() {
             self.bot
-                .chat(&format!("/w {} {}", self.chat.sender().unwrap(), message));
+                .chat(&format!("/w {} {message}", self.chat.sender().unwrap()));
         } else {
-            self.bot.chat(message);
+            self.bot.chat(&message);
         }
     }